Jump animations from new travel system missing on native Mac client https://jira.ccpgames.com/browse/EO-12702
Fixing a few other instances of potentially variable-size int variables exposed to blue.
Doing this in mainline because the change is safe and because the change is touching too many files and the validation for these issues is in mainline

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1879815]

+++ Tr2DynamicBinding.h
@@ -47,7 +47,7 @@ private:
 	BlueSharedString m_sourceObjectAttribute;
 	BlueWeakRef<IRoot> m_source;
 	float m_scale;
-	long m_bindingDelay;
+	int32_t m_bindingDelay;
 
 	Be::Time m_bindingTime;
